+++ custodia/document.go
@@ -4,8 +4,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"strings"
-	"time"
+	"net/url"
 
 	"github.com/google/uuid"
 	"github.com/simplereach/timeutils"
@@ -151,62 +150,30 @@ func (ca *CustodiaAPIv1) DeleteDocument(documentId uuid.UUID, force,
 }
 
 // [L]ist all the documents in a Schema
-// url params:
+// url queryParams:
+//   offset: int: number of items to skip from the beginning of the list
+//   limit: int : maximum number of items to return in a single page
 //   full_document: bool
 //   is_active: bool
-//   insert_date__gt: time.Time
-//   insert_date__lt: time.Time
-//   last_update__gt: time.Time
-//   last_update__lt: time.Time
+//   insert_date__gt: time string (RFC3339, YYYY-MM-DDTHH:MM:SS)
+//   insert_date__lt: time string (RFC3339)
+//   last_update__gt: time string (RFC3339)
+//   last_update__lt: time string (RFC3339)
 func (ca *CustodiaAPIv1) ListDocuments(schema Schema,
-	params map[string]any) ([]*Document, error) {
-	url := fmt.Sprintf("/schemas/%s/documents", schema.Id)
-	if len(params) > 0 {
-		url += "?"
+	queryParams map[string]string) ([]*Document, error) {
+	u, err := url.Parse(fmt.Sprintf("/schemas/%s/documents", schema.Id))
+	if err != nil {
+		return nil, fmt.Errorf("error parsing url: %v", err)
 	}
 
-	availableParams := map[string]any{
-		"full_document": true,
-		"is_active": true,
-		"insert_date__gt": time.Time{},
-		"insert_date__lt": time.Time{},
-		"last_update__gt": time.Time{},
-		"last_update__lt": time.Time{},
+	// Adding query params
+	q := u.Query()
+	for k, v := range(queryParams) {
+		q.Set(k, v)
 	}
-	for param := range params {
-		// check that param is legit
-		_, ok := availableParams[param]
-		if !ok {
-			return nil, fmt.Errorf("got unexpected param '%s'", param)
-		}
-		value := params[param]
+	u.RawQuery = q.Encode()
 
-		switch param {
-		case "full_document", "is_active":
-			_, ok := value.(bool)
-			if !ok {
-				err := fmt.Errorf("param '%s': bad type: '%T', must be bool",
-					param, value)
-				return nil, err
-			}
-		case "insert_date__gt", "insert_date__lt", "last_update__gt",
-			"last_update__lt":
-			time_value, ok := value.(time.Time)
-			if !ok {
-				err := fmt.Errorf("param '%s': bad type: '%T', must be Time",
-					param, value)
-				return nil, err
-			}
-			value = time_value.Format(time.RFC3339)
-		default:
-			return nil, fmt.Errorf("got unexpected param '%s'", param)
-		}
-
-		url += fmt.Sprintf("%s=%v&", param, value)
-	}
-
-	url = strings.TrimRight(url, "&")
-	resp, err := ca.Call("GET", url, nil)
+	resp, err := ca.Call("GET", u.String(), nil)
 	if err != nil {
 		return nil, err
 	}